+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Adds [i18n rewrites](https://firebase.google.com/docs/hosting/i18n-rewrites) to the Firebase Hosting emulator.
+- Changes Auth Emulator IDP sign-in to always return an access token.